bin/font-merging/requirements.txt
fonttools
uharfbuzz
openpyxl
pillow
freetype-py